LayoutTests/imported/w3c/web-platform-tests/content-security-policy/unsafe-hashes/script_event_handlers_denied_missing_unsafe_hashes-expected.txt


FAIL Test that the inline event handler is not allowed to run assert_equals: expected "script-src-attr" but got "script-src"
